requirements.txt
Flask==2.2.2
flask-jwt-extended==4.4.4
psycopg2-binary==2.9.6
requests==2.28.1